Mv log filtering to log pkg
