+++ pkg/flow/grpc-logs_test.go
@@ -5,7 +5,6 @@ import (
 	_ "embed"
 	"encoding/json"
 	"fmt"
-	"strings"
 	"testing"
 
 	"github.com/direktiv/direktiv/pkg/flow/grpc"
@@ -17,12 +16,6 @@ import (
 
 //go:embed mockdata/entlog_loop.json
 var loopjson string
-
-//go:embed mockdata/entlog_loopFunctionNested.json
-var loopnestedjson string
-
-//go:embed mockdata/entlog_nestedLoop.json
-var loopnestedloopjson string
 
 const loopJsonValidInstanceID = "1a0d5909-223f-4f44-86d7-1833ab4d21c8"
 
@@ -38,134 +31,6 @@ var expectedLoopnestedjsonWFLooperlooperSTATESolveIt2 = []string{
 	"Running state logic",
 	"returned",
 	"returned",
-}
-
-func TestQueryMatchState(t *testing.T) {
-	res := assertQueryMatchState(t, loopjson, "test", "", "", 21)
-	assertsortedByTime(t, res)
-	res = assertQueryMatchState(t, loopjson, "test", "solve", "", 16)
-	assertsortedByTime(t, res)
-	res = assertQueryMatchState(t, loopjson, "test", "solve", "1", 6)
-	assertsortedByTime(t, res)
-	res = assertQueryMatchState(t, loopnestedjson, "looperlooper", "solve", "", 12)
-	assertsortedByTime(t, res)
-	res = assertQueryMatchState(t, loopnestedjson, "looper", "", "3", 8)
-	assertsortedByTime(t, res)
-	res = assertQueryMatchState(t, loopnestedjson, "looper", "", "3", 8)
-	assertsortedByTime(t, res)
-	res = assertQueryMatchState(t, loopnestedloopjson, "looperlooper", "solve", "0", 51)
-	assertsortedByTime(t, res)
-	res = assertQueryMatchState(t, loopnestedjson, "looperlooper", "solve", "0", 10)
-	assertLogmsgsContain(t, res, expectedLoopnestedjsonWFLooperlooperSTATESolveIt2)
-	assertLogmsgsPresent(t, res, expectedLoopnestedjsonWFLooperlooperSTATESolveIt2)
-	res = assertQueryMatchState(t, loopnestedloopjson, "looper", "solve", "0", 13)
-	assertsortedByTime(t, res)
-	res = assertQueryMatchState(t, loopjson, "test", "", "100", 0)
-	assertsortedByTime(t, res)
-}
-
-func assertQueryMatchState(t *testing.T, jsondump, wf, state, iterator string, resLen int) []*internallogger.LogMsgs {
-	t.Helper()
-	logmsgs := make([]*internallogger.LogMsgs, 0)
-	err := json.Unmarshal([]byte(jsondump), &logmsgs)
-	if err != nil {
-		t.Error(err)
-	}
-	res := filterMatchByWfStateIterator(wf+"::"+state+"::"+iterator, logmsgs)
-	if len(res) != resLen {
-		t.Errorf("len off; was %d, want %d", len(res), resLen)
-	}
-	for _, v := range res {
-		nestedloopPresent := checkNestedLoop(res)
-		if iterator == "" && state != "" {
-			assertTagsFiltered(t, v.Tags, "workflow", wf)
-			assertTagsFiltered(t, v.Tags, "state-id", state)
-		}
-		if iterator == "" && state == "" {
-			assertTagsFiltered(t, v.Tags, "workflow", wf)
-		}
-		if !nestedloopPresent && iterator != "" {
-			assertTagsFiltered(t, v.Tags, "loop-index", iterator)
-		}
-	}
-	resSecond := filterMatchByWfStateIterator(wf+"::"+state+"::"+iterator, res)
-	if len(res) != len(resSecond) {
-		t.Errorf("len off when runned second time; was first run %d, is %d, should %d", len(res), len(resSecond), resLen)
-	}
-	return res
-}
-
-func TestWhiteboxTestServerLogs(t *testing.T) {
-	srv := server{}
-	flowSrv := flow{}
-
-	flowSrv.server = &srv
-	logs, logobserver := testutils.ObservedLogger()
-	gdb, cleanup, err := testutils.DatabaseGorm()
-	if err != nil {
-		t.Error(err)
-	}
-	defer func() {
-		err := cleanup()
-		if err != nil {
-			fmt.Println(err)
-		}
-	}()
-	srv.gormDB = gdb
-	srv.sugar = logs
-	reqSrvLogs := grpc.ServerLogsRequest{
-		Pagination: &grpc.Pagination{},
-	}
-	resSrvLogs := requestServerLogs(t, flowSrv, &reqSrvLogs)
-	if len(logobserver.All()) <= 0 {
-		t.Error("some logmsg should heve been printed")
-	}
-	if int(resSrvLogs.PageInfo.Limit) > len(resSrvLogs.Results) {
-		t.Errorf("got more results then specified in pageinfo")
-	}
-}
-
-func TestFilterByIterrator(t *testing.T) {
-	logmsgs := make([]*internallogger.LogMsgs, 0)
-	err := json.Unmarshal([]byte(loopnestedjson), &logmsgs)
-	if err != nil {
-		t.Error(err)
-		return
-	}
-	child := filterByIterrator("2", logmsgs)
-	if child == nil {
-		t.Errorf("did not found")
-	}
-	res := filterByIterrator("", logmsgs)
-	if len(res) != 0 {
-		t.Errorf("calling filterByIterrator with empty iterator is a invalid operation and should result a empty slice")
-	}
-}
-
-func TestFilterLogmsg(t *testing.T) {
-	logmsgs := make([]*internallogger.LogMsgs, 0)
-	err := json.Unmarshal([]byte(loopnestedjson), &logmsgs)
-	if err != nil {
-		t.Error(err)
-		return
-	}
-	res := filterLogmsg(&grpc.PageFilter{
-		Field: "TESTFIELD",
-		Type:  "TESTTYPE",
-	}, logmsgs)
-	ok := assertEquals(t, logmsgs, res)
-	if !ok {
-		t.Error("input slice should not be modified if Pagefilter does not match")
-	}
-	res = filterLogmsg(&grpc.PageFilter{
-		Field: "QUERY",
-		Type:  "MATCH",
-		Val:   "looperlooper::solve",
-	}, logmsgs)
-	equals := assertEquals(t, logmsgs, res)
-	if equals {
-		t.Error("input slice should have been filtered")
-	}
 }
 
 func requestServerLogs(t *testing.T, flowSrv flow, req *grpc.ServerLogsRequest) *grpc.ServerLogsResponse {
@@ -304,82 +169,6 @@ func assertNoMissingLogs(t *testing.T, res *grpc.InstanceLogsResponse, in map[st
 	}
 }
 
-func assertLogmsgsContain(t *testing.T, msgs []*internallogger.LogMsgs, expectedMsg []string) {
-	t.Helper()
-	for _, v := range msgs {
-		ok := false
-		for _, e := range expectedMsg {
-			ok = ok || strings.Contains(v.Msg, e)
-		}
-		if !ok {
-			t.Errorf("logmsg %s was not expected", v.Msg)
-		}
-	}
-}
-
-func assertEquals(t *testing.T, a []*internallogger.LogMsgs, b []*internallogger.LogMsgs) bool {
-	t.Helper()
-	if len(a) != len(b) {
-		return false
-	}
-	for i, v := range a {
-		if v != b[i] {
-			return false
-		}
-	}
-	return true
-}
-
-func assertLogmsgsPresent(t *testing.T, msgs []*internallogger.LogMsgs, expectedMsg []string) {
-	t.Helper()
-
-	for _, e := range expectedMsg {
-		ok := false
-		for _, v := range msgs {
-			ok = ok || strings.Contains(v.Msg, e)
-		}
-		if !ok {
-			t.Errorf("logmsg %s was missing", e)
-		}
-	}
-}
-
-func assertTagsFiltered(t *testing.T, tags map[string]interface{}, tag, value string) {
-	t.Helper()
-	if tags[tag] != value {
-		t.Errorf("found wrong tag-value for %s: %s want %s", tag, tags[tag], value)
-	}
-}
-
-func assertsortedByTime(t *testing.T, in []*internallogger.LogMsgs) bool {
-	t.Helper()
-	if len(in) < 2 {
-		return true
-	}
-
-	for i := 2; i < len(in); i = i + 2 {
-		a := in[i]
-		b := in[i]
-		if i+1 < len(in) {
-			b = in[i+1]
-		}
-		if a.T.After(b.T) {
-			t.Errorf("Array not sorted")
-			return false
-		}
-	}
-	return true
-}
-
-func checkNestedLoop(in []*internallogger.LogMsgs) bool {
-	for _, v := range in {
-		if v.Tags["state-type"] == "foreach" {
-			return true
-		}
-	}
-	return false
-}
-
 func storeLogmsg(ctx context.Context, db *gorm.DB, l *internallogger.LogMsgs) (*internallogger.LogMsgs, error) {
 	l.Oid = uuid.New()
 	db = db.Debug()
@@ -390,3 +179,33 @@ func storeLogmsg(ctx context.Context, db *gorm.DB, l *internallogger.LogMsgs) (*
 	}
 	return l, nil
 }
+
+// func TestWhiteboxTestServerLogs(t *testing.T) {
+// 	srv := server{}
+// 	flowSrv := flow{}
+
+// 	flowSrv.server = &srv
+// 	logs, logobserver := testutils.ObservedLogger()
+// 	gdb, cleanup, err := testutils.DatabaseGorm()
+// 	if err != nil {
+// 		t.Error(err)
+// 	}
+// 	defer func() {
+// 		err := cleanup()
+// 		if err != nil {
+// 			fmt.Println(err)
+// 		}
+// 	}()
+// 	srv.gormDB = gdb
+// 	srv.sugar = logs
+// 	reqSrvLogs := grpc.ServerLogsRequest{
+// 		Pagination: &grpc.Pagination{},
+// 	}
+// 	resSrvLogs := requestServerLogs(t, flowSrv, &reqSrvLogs)
+// 	if len(logobserver.All()) <= 0 {
+// 		t.Error("some logmsg should heve been printed")
+// 	}
+// 	if int(resSrvLogs.PageInfo.Limit) > len(resSrvLogs.Results) {
+// 		t.Errorf("got more results then specified in pageinfo")
+// 	}
+// }
