+++ pkg/flow/internallogger/logs_test.go
@@ -2,16 +2,45 @@ package internallogger
 
 import (
 	"context"
+	_ "embed"
+	"encoding/json"
 	"fmt"
+	"strings"
 	"testing"
 	"time"
 
 	"github.com/direktiv/direktiv/pkg/flow/database/recipient"
+	"github.com/direktiv/direktiv/pkg/flow/grpc"
 	"github.com/direktiv/direktiv/pkg/flow/internal/testutils"
 	"github.com/google/uuid"
 )
 
+//go:embed mockdata/entlog_loop.json
+var loopjson string
+
+//go:embed mockdata/entlog_loopFunctionNested.json
+var loopnestedjson string
+
+//go:embed mockdata/entlog_nestedLoop.json
+var loopnestedloopjson string
+
 var _notifyLogsTriggeredWith notifyLogsTriggeredWith
+
+const loopJsonValidInstanceID = "1a0d5909-223f-4f44-86d7-1833ab4d21c8"
+
+var expectedLoopnestedjsonWFLooperlooperSTATESolveIt2 = []string{
+	"Preparing workflow triggered by",
+	"Sleeping until",
+	"Starting workflow",
+	"Running state logic",
+	"Creating cookie",
+	"Creating new",
+	"Creating URL",
+	"Sending request",
+	"Running state logic",
+	"returned",
+	"returned",
+}
 
 func TestStoringLogMsg(t *testing.T) {
 	sugar, telemetrylogs := testutils.ObservedLogger()
@@ -75,7 +104,7 @@ func TestTelemetry(t *testing.T) {
 	tags := make(map[string]string)
 	tags["recipientType"] = recipient.Server.String()
 	msg := "test3"
-	logger.Telemetry(context.Background(), Info, nil, msg)
+	logger.telemetry(context.Background(), Info, nil, msg)
 	if len(telemetrylogs.All()) < 1 {
 		t.Errorf("len of telemetry logs wrong")
 		return
@@ -93,7 +122,7 @@ func TestTelemetryWithTags(t *testing.T) {
 	tags := make(map[string]string)
 	tags["recipientType"] = recipient.Server.String()
 	msg := "test4"
-	logger.Telemetry(context.Background(), Info, tags, msg)
+	logger.telemetry(context.Background(), Info, tags, msg)
 	if len(telemetrylogs.All()) < 1 {
 		t.Errorf("len of telemetry logs wrong")
 		return
@@ -122,7 +151,7 @@ func TestSendLogMsgToDB(t *testing.T) {
 	tags["recipientType"] = recipient.Server.String()
 	recipent := uuid.New()
 	msg := "test1"
-	err = logger.SendLogMsgToDB(&logMsg{
+	err = logger.sendLogMsgToDB(&logMsg{
 		recipientID:   recipent,
 		recipientType: recipient.Instance,
 		LogMsgs: &LogMsgs{
@@ -179,4 +208,178 @@ func (ln *LogNotifyMock) NotifyLogs(recipientID uuid.UUID, recipientType recipie
 type notifyLogsTriggeredWith struct {
 	Id            uuid.UUID
 	RecipientType recipient.RecipientType
+}
+
+func TestQueryMatchState(t *testing.T) {
+	res := assertQueryMatchState(t, loopjson, "test", "", "", 21)
+	assertsortedByTime(t, res)
+	res = assertQueryMatchState(t, loopjson, "test", "solve", "", 16)
+	assertsortedByTime(t, res)
+	res = assertQueryMatchState(t, loopjson, "test", "solve", "1", 6)
+	assertsortedByTime(t, res)
+	res = assertQueryMatchState(t, loopnestedjson, "looperlooper", "solve", "", 12)
+	assertsortedByTime(t, res)
+	res = assertQueryMatchState(t, loopnestedjson, "looper", "", "3", 8)
+	assertsortedByTime(t, res)
+	res = assertQueryMatchState(t, loopnestedjson, "looper", "", "3", 8)
+	assertsortedByTime(t, res)
+	res = assertQueryMatchState(t, loopnestedloopjson, "looperlooper", "solve", "0", 51)
+	assertsortedByTime(t, res)
+	res = assertQueryMatchState(t, loopnestedjson, "looperlooper", "solve", "0", 10)
+	assertLogmsgsContain(t, res, expectedLoopnestedjsonWFLooperlooperSTATESolveIt2)
+	assertLogmsgsPresent(t, res, expectedLoopnestedjsonWFLooperlooperSTATESolveIt2)
+	res = assertQueryMatchState(t, loopnestedloopjson, "looper", "solve", "0", 13)
+	assertsortedByTime(t, res)
+	res = assertQueryMatchState(t, loopjson, "test", "", "100", 0)
+	assertsortedByTime(t, res)
+}
+
+func assertLogmsgsPresent(t *testing.T, msgs []*LogMsgs, expectedMsg []string) {
+	t.Helper()
+
+	for _, e := range expectedMsg {
+		ok := false
+		for _, v := range msgs {
+			ok = ok || strings.Contains(v.Msg, e)
+		}
+		if !ok {
+			t.Errorf("logmsg %s was missing", e)
+		}
+	}
+}
+
+func assertLogmsgsContain(t *testing.T, msgs []*LogMsgs, expectedMsg []string) {
+	t.Helper()
+	for _, v := range msgs {
+		ok := false
+		for _, e := range expectedMsg {
+			ok = ok || strings.Contains(v.Msg, e)
+		}
+		if !ok {
+			t.Errorf("logmsg %s was not expected", v.Msg)
+		}
+	}
+}
+
+func assertsortedByTime(t *testing.T, in []*LogMsgs) bool {
+	t.Helper()
+	if len(in) < 2 {
+		return true
+	}
+
+	for i := 2; i < len(in); i = i + 2 {
+		a := in[i]
+		b := in[i]
+		if i+1 < len(in) {
+			b = in[i+1]
+		}
+		if a.T.After(b.T) {
+			t.Errorf("Array not sorted")
+			return false
+		}
+	}
+	return true
+}
+
+func assertQueryMatchState(t *testing.T, jsondump, wf, state, iterator string, resLen int) []*LogMsgs {
+	t.Helper()
+	logmsgs := make([]*LogMsgs, 0)
+	err := json.Unmarshal([]byte(jsondump), &logmsgs)
+	if err != nil {
+		t.Error(err)
+	}
+	res := filterMatchByWfStateIterator(wf+"::"+state+"::"+iterator, logmsgs)
+	if len(res) != resLen {
+		t.Errorf("len off; was %d, want %d", len(res), resLen)
+	}
+	for _, v := range res {
+		nestedloopPresent := checkNestedLoop(res)
+		if iterator == "" && state != "" {
+			assertTagsFiltered(t, v.Tags, "workflow", wf)
+			assertTagsFiltered(t, v.Tags, "state-id", state)
+		}
+		if iterator == "" && state == "" {
+			assertTagsFiltered(t, v.Tags, "workflow", wf)
+		}
+		if !nestedloopPresent && iterator != "" {
+			assertTagsFiltered(t, v.Tags, "loop-index", iterator)
+		}
+	}
+	resSecond := filterMatchByWfStateIterator(wf+"::"+state+"::"+iterator, res)
+	if len(res) != len(resSecond) {
+		t.Errorf("len off when runned second time; was first run %d, is %d, should %d", len(res), len(resSecond), resLen)
+	}
+	return res
+}
+
+func checkNestedLoop(in []*LogMsgs) bool {
+	for _, v := range in {
+		if v.Tags["state-type"] == "foreach" {
+			return true
+		}
+	}
+	return false
+}
+
+func assertTagsFiltered(t *testing.T, tags map[string]interface{}, tag, value string) {
+	t.Helper()
+	if tags[tag] != value {
+		t.Errorf("found wrong tag-value for %s: %s want %s", tag, tags[tag], value)
+	}
+}
+
+func TestFilterByIterrator(t *testing.T) {
+	logmsgs := make([]*LogMsgs, 0)
+	err := json.Unmarshal([]byte(loopnestedjson), &logmsgs)
+	if err != nil {
+		t.Error(err)
+		return
+	}
+	child := filterByIterrator("2", logmsgs)
+	if child == nil {
+		t.Errorf("did not found")
+	}
+	res := filterByIterrator("", logmsgs)
+	if len(res) != 0 {
+		t.Errorf("calling filterByIterrator with empty iterator is a invalid operation and should result a empty slice")
+	}
+}
+
+func TestFilterLogmsg(t *testing.T) {
+	logmsgs := make([]*LogMsgs, 0)
+	err := json.Unmarshal([]byte(loopnestedjson), &logmsgs)
+	if err != nil {
+		t.Error(err)
+		return
+	}
+	res := FilterLogmsg(&grpc.PageFilter{
+		Field: "TESTFIELD",
+		Type:  "TESTTYPE",
+	}, logmsgs)
+	ok := assertEquals(t, logmsgs, res)
+	if !ok {
+		t.Error("input slice should not be modified if Pagefilter does not match")
+	}
+	res = FilterLogmsg(&grpc.PageFilter{
+		Field: "QUERY",
+		Type:  "MATCH",
+		Val:   "looperlooper::solve",
+	}, logmsgs)
+	equals := assertEquals(t, logmsgs, res)
+	if equals {
+		t.Error("input slice should have been filtered")
+	}
+}
+
+func assertEquals(t *testing.T, a []*LogMsgs, b []*LogMsgs) bool {
+	t.Helper()
+	if len(a) != len(b) {
+		return false
+	}
+	for i, v := range a {
+		if v != b[i] {
+			return false
+		}
+	}
+	return true
 }
